+++ linknamu/src/main/java/com/kakao/linknamu/tag/service/TagDeleteService.java
@@ -1,25 +1,23 @@
 package com.kakao.linknamu.tag.service;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.kakao.linknamu.core.exception.Exception404;
 import com.kakao.linknamu.tag.TagExceptionStatus;
 import com.kakao.linknamu.tag.entity.Tag;
-import com.kakao.linknamu.tag.repository.TagJPARepository;
-
+import com.kakao.linknamu.tag.repository.TagJpaRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
 @RequiredArgsConstructor
 @Service
 public class TagDeleteService {
-	private final TagJPARepository tagJPARepository;
+	private final TagJpaRepository tagJpaRepository;
 
 	public void deleteTagByName(Long userId, String name) {
-		Tag tag = tagJPARepository.findByUserIdAndName(userId, name).orElseThrow(
+		Tag tag = tagJpaRepository.findByUserIdAndName(userId, name).orElseThrow(
 			() -> new Exception404(TagExceptionStatus.TAG_NOT_FOUND)
 		);
-		tagJPARepository.delete(tag);
+		tagJpaRepository.delete(tag);
 	}
 }
