+++ linknamu/src/main/java/com/kakao/linknamu/tag/service/TagSaveService.java
@@ -1,7 +1,7 @@
 package com.kakao.linknamu.tag.service;
 
 import com.kakao.linknamu.tag.entity.Tag;
-import com.kakao.linknamu.tag.repository.TagJPARepository;
+import com.kakao.linknamu.tag.repository.TagJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,9 +10,9 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Service
 public class TagSaveService {
-    private final TagJPARepository tagJPARepository;
+	private final TagJpaRepository tagJpaRepository;
 
-    public void createTag(Tag newTag) {
-        tagJPARepository.save(newTag);
-    }
+	public void createTag(Tag newTag) {
+		tagJpaRepository.save(newTag);
+	}
 }
